+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes
 
 - 1.7.2
     - BUGFIX: optically better grey out meter and sliders on login
+    - IMPROVEMENT: optically simplify the login dialog
 
 - 1.7.1
     - EXTENSION: verify digital signature of distribution archive on all updates